refactor: break generate_cluster_metadata file into individual files for each serializer type
